DIV-5908: Moving the remaining mappers from CFS into the div-common-lib
