--- src/main/java/uk/gov/hmcts/reform/divorce/caseformatterservice/domain/model/ccd/DaCaseData.java
@@ -1,13 +1,0 @@
-package uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.ccd;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Data
-public class DaCaseData {
-
-    @JsonProperty("ApplyForDecreeAbsolute")
-    private String applyForDecreeAbsolute;
-}
