+++ src/main/java/uk/gov/hmcts/reform/divorce/caseformatterservice/domain/model/documentupdate/DocumentUpdateRequest.java
@@ -1,6 +1,7 @@
 package uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.documentupdate;
 
 import lombok.Data;
+import uk.gov.hmcts.reform.divorce.model.documentupdate.GeneratedDocumentInfo;
 
 import java.util.List;
 import java.util.Map;
