--- src/main/java/uk/gov/hmcts/reform/divorce/caseformatterservice/domain/model/documentupdate/GeneratedDocumentInfo.java
@@ -1,14 +1,0 @@
-package uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.documentupdate;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Builder
-public class GeneratedDocumentInfo {
-    private String url;
-    private String documentType;
-    private String fileName;
-}
