--- src/test/resources/fixtures/divorcetoccdmapping/ccd/dn-clarification-existing-data.json
@@ -1,70 +1,0 @@
-{
-    "DnClarificationResponse" : [
-        {
-            "id" : "initial-id",
-            "value" : "Clarification 1: This is the initial response"
-        },
-        {
-            "id" : null,
-            "value" : "Clarification 2: This is a response attempt."
-        }
-    ],
-    "DnClarificationUploadDocuments" : [
-        {
-            "id" : "initial-id",
-            "value" : "Clarification 1: No"
-        },
-        {
-            "id" : null,
-            "value" : "Clarification 2: Yes"
-        }
-    ],
-    "DocumentsUploadedDnClarification" : [
-        {
-            "id" : null,
-            "value" : {
-                "DocumentType" : "other",
-                "DocumentLink" : {
-                    "document_url" : "http://localhost:5006/documents/1234567-abcd-1234-wxyz-098765432101",
-                    "document_binary_url" : null,
-                    "document_filename" : null
-                },
-                "DocumentDateAdded" : "2011-11-11",
-                "DocumentComment" : "Existing Comment",
-                "DocumentFileName" : "favicon.ico",
-                "DocumentEmailContent" : ""
-            }
-        },
-        {
-            "id" : null,
-            "value" : {
-                "DocumentType" : "other",
-                "DocumentLink" : {
-                    "document_url" : "http://localhost:5006/documents/7f63ca9b-c361-49ab-aa8c-8fbdb6bc2936",
-                    "document_binary_url" : null,
-                    "document_filename" : null
-                },
-                "DocumentDateAdded" : "2017-12-11",
-                "DocumentComment" : "Clarification 2: Document",
-                "DocumentFileName" : "govuklogo.png",
-                "DocumentEmailContent" : ""
-            }
-        },
-        {
-            "id" : null,
-            "value" : {
-                "DocumentType": "other",
-                "DocumentLink" : {
-                    "document_url" : "http://localhost:5006/documents/560f4bd1-fcf2-4dc5-80d1-d7e00d861655",
-                    "document_binary_url" : null,
-                    "document_filename" : null
-                },
-                "DocumentDateAdded": "2017-12-11",
-                "DocumentComment": "Clarification 2: Document",
-                "DocumentFileName": "d8-eng.pdf",
-                "DocumentEmailContent": ""
-            }
-        }
-    ],
-    "ClarificationDigital": "YES"
-}
