--- src/test/resources/fixtures/divorcetoccdmapping/ccd/dn-null.json
@@ -1,25 +1,0 @@
-{
-    "PetitionChangedYesNoDN" : "YES",
-    "PetitionChangedDetailsDN" : "Some details about this dn changes",
-    "ConfirmPetitionDN" : "YES",
-    "DivorceCostsOptionDN" : "differentAmount",
-    "CostsDifferentDetails" : "I want to pay 60%",
-    "statementOfTruthDN" : "YES",
-    "AdulteryLifeIntolerable" : null,
-    "AdulteryDateFoundOut" : null,
-    "AdulteryLivedApartSinceEventDN" : null,
-    "AdulteryTimeLivedTogetherDetailsDN" : null,
-    "BehaviourStillHappeningDN" : "NO",
-    "BehaviourMostRecentIncidentDateDN" : "2018-10-03",
-    "BehaviourLivedApartSinceEventDN" : "NO",
-    "BehaviourTimeLivedTogetherDetailsDN" : "Since we've met in a bar",
-    "DesertionLivedApartSinceEventDN" : "NO",
-    "DesertionTimeLivedTogetherDetailsDN" : "Since we've met in a bar",
-    "SeparationLivedApartSinceEventDN" : "NO",
-    "SeparationTimeLivedTogetherDetailsDN" : "Since we've met in a bar",
-    "DesertionAskedToResumeDN" : "YES",
-    "DesertionAskedToResumeDNRefused" : "YES",
-    "DesertionAskedToResumeDNDetails" : "Refused to stay",
-    "DocumentsUploadedDN" : null,
-    "DocumentsUploadedQuestionDN" : "YES"
-}
