--- src/test/resources/fixtures/divorcetoccdmapping/divorce/dn-clarification.json
@@ -1,27 +1,0 @@
-{
-    "clarificationResponse": "This is a response attempt.",
-    "uploadAnyOtherDocuments": "Yes",
-    "files" : [
-        {
-            "createdBy" : 12661,
-            "createdOn" : "2017-12-11",
-            "lastModifiedBy" : 12661,
-            "modifiedOn" : "2017-12-11",
-            "fileName" : "govuklogo.png",
-            "fileUrl" : "http://localhost:5006/documents/7f63ca9b-c361-49ab-aa8c-8fbdb6bc2936",
-            "mimeType" : "image/png",
-            "status" : "OK"
-        },
-        {
-            "createdBy" : 12661,
-            "createdOn" : "2017-12-11",
-            "lastModifiedBy" : 12661,
-            "modifiedOn" : "2017-12-11",
-            "fileName" : "d8-eng.pdf",
-            "fileUrl" : "http://localhost:5006/documents/560f4bd1-fcf2-4dc5-80d1-d7e00d861655",
-            "mimeType" : "application/pdf",
-            "status" : "OK"
-        }
-    ],
-    "clarificationDigital": "Yes"
-}
