--- src/test/resources/fixtures/model/divorce/co-respondent/co-respondent-answers-no-cost.json
@@ -1,14 +1,0 @@
-{
-    "coRespondentAnswers": {
-        "contactInfo": {
-            "emailAddress": "co-respondent@divorce.co.uk",
-            "consentToReceivingEmails": "Yes",
-            "contactMethodIsDigital": "No",
-            "phoneNumber": "+4404839202839"
-        },
-        "confirmReadPetition": "No",
-        "statementOfTruth": "Yes",
-        "admitAdultery": "Yes",
-        "defendsDivorce": "No"
-    }
-}
